Change messageUID from in to long
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@1102 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-ejb/src/com/liferay/portlet/mailbox/action/MailboxAction.java
@@ -132,7 +132,7 @@ public class MailboxAction extends JSONAction {
 		JSONObject jsonObj = new JSONObject();
 
 		String folderId = ParamUtil.getString(req, "folderId");
-		int messageId = ParamUtil.getInteger(req, "messageId");
+		long messageId = ParamUtil.getLong(req, "messageId");
 
 		MailUtil.setCurrentFolder(req.getSession(), folderId);
